split up draft into files
