+++ webhooks/src/main/scala/zio/webhooks/NonEmptySet.scala
@@ -1,0 +1,3 @@
+package zio.webhooks
+
+final case class NonEmptySet[A](head: A, others: Set[A])
