+++ webhooks/src/main/scala/zio/webhooks/WebhookDeliveryBatching.scala
@@ -1,0 +1,7 @@
+package zio.webhooks
+
+sealed trait WebhookDeliveryBatching
+object WebhookDeliveryBatching {
+  case object Single  extends WebhookDeliveryBatching
+  case object Batched extends WebhookDeliveryBatching
+}
