+++ webhooks/src/main/scala/zio/webhooks/WebhookDeliverySemantics.scala
@@ -1,0 +1,7 @@
+package zio.webhooks
+
+sealed trait WebhookDeliverySemantics
+object WebhookDeliverySemantics {
+  case object AtLeastOnce extends WebhookDeliverySemantics
+  case object AtMostOnce  extends WebhookDeliverySemantics
+}
