+++ webhooks/src/main/scala/zio/webhooks/WebhookEventId.scala
@@ -1,0 +1,3 @@
+package zio.webhooks
+
+final case class WebhookEventId(value: Long)
