+++ webhooks/src/main/scala/zio/webhooks/WebhookId.scala
@@ -1,0 +1,6 @@
+package zio.webhooks
+
+/**
+ * A webhook is simply identified by a [[Long]].
+ */
+final case class WebhookId(value: Long)
