+++ webhooks/src/main/scala/zio/webhooks/WebhookKey.scala
@@ -1,0 +1,3 @@
+package zio.webhooks
+
+final case class WebhookKey(eventId: WebhookEventId, webhookId: WebhookId)
